✨ Disable Default Emoji Usage
Set the default value for `use_emoji` configuration to `false`.

This change modifies the default behavior of the `use_emoji` setting in the
configuration to be disabled.
Previously, the default value was `true`, meaning emojis were enabled by
default.
This change ensures that emojis are only used if explicitly enabled by the
user in the configuration.